refactor special section using css module
