+++ src/Components/Main/Special/MenuCard/Delivery.js
@@ -1,6 +1,8 @@
+import styles from "./MenuCard.module.css";
+
 export default function Delivery() {
   return (
-    <button className="delivery-btn">
+    <button className={styles.deliveryBtn}>
       Order a delivery <img src="./assets/motorbike.svg" alt="motorbike" />
     </button>
   );
