+++ src/Components/Main/Special/MenuCard/MenuCard.module.css
@@ -1,0 +1,69 @@
+@import "../../../../App.css";
+
+.cardWrapper {
+  /* width: 265px; */
+  min-height: 31.25rem;
+  padding-bottom: 2.5rem;
+  display: flex;
+  flex-direction: column;
+  border-radius: 1rem;
+  overflow: hidden;
+  background-color: var(--highlight-lite);
+  position: relative;
+}
+
+.photo {
+  width: 100%;
+  height: 12.5rem;
+  object-fit: cover;
+  margin-bottom: 0.9375rem;
+  transition: all ease-in-out 0.3s;
+}
+
+.photo:hover {
+  transform: scale(1.1);
+}
+
+.price {
+  color: var(--secondary-orange);
+  font-weight: 600;
+}
+
+.price::before {
+  content: "$";
+  padding-right: 0.5rem;
+}
+
+.name {
+  font-family: "Markazi Text";
+  font-size: 24px;
+}
+
+.description {
+  color: var(--primary-green);
+  line-height: auto;
+}
+
+.deliveryBtnWrapper {
+  position: absolute;
+  bottom: 1.5rem;
+}
+
+.deliveryBtn {
+  border: none;
+  font-weight: bold;
+  font-size: 16px;
+  border-radius: 1px;
+  color: var(--highlight-dark);
+  cursor: pointer;
+  transition: transform 0.3s;
+}
+
+.deliveryBtn:active {
+  transform: scale(0.9);
+}
+
+.deliveryBtn:hover {
+  outline: 1px solid var(--primary-green);
+  outline-offset: 10px;
+}
