+++ src/Components/Main/Special/Special.module.css
@@ -1,0 +1,7 @@
+@import "../../../App.css";
+
+.menuCardsWrapper {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 2rem;
+}
